requirements.txt
# requirements.txt

tensorflow>=2.0.0
pymongo>=3.13.0
pandas>=1.3.3
numpy>=1.21.2
PyYAML>=5.4.1
